refactor: move ErrorBoundary into root.tsx file
