--- styled-components/app/routes/_boundary.tsx
@@ -1,33 +1,0 @@
-import { isRouteErrorResponse, useRouteError } from "@remix-run/react";
-
-import { Box } from "~/components/Box";
-
-export function ErrorBoundary() {
-    const error = useRouteError();
-
-    if (isRouteErrorResponse(error)) {
-        return (
-            <Box>
-              <h1>Catch Boundary</h1>
-              <p>
-                {error.status} {error.statusText}
-              </p>
-            </Box>
-          );
-    }
-
-    let errorMessage = "Unknown error";
-    let errorStatus = 500;
-    if (error instanceof Error) {
-        errorMessage = error.message;
-    }
-
-  return (
-    <Box>
-        <h1>Error Boundary</h1>
-        <p>
-            {errorStatus} {errorMessage}
-        </p>
-    </Box>
-  );
-}
